+++ backend/core/infrastructure/factories/abstract_factories.py
@@ -1,15 +1,20 @@
 import pandas as pd
 from backend.core.adapters.impl_bias_metrics import ImplBiasMetrics
 from backend.core.adapters.impl_file_converter import ImplFileConverter
-from backend.core.data_access.file_repository import FileRepository
-from backend.core.use_cases.interfaces import BiasMetrics, FileConverter
+from backend.core.data_access.LocalFileRepo import LocalFileRepo
+from backend.core.data_access.S3FileRepo import S3FileRepo
+from backend.core.use_cases.interfaces import BiasMetrics, FileConverter, FileRepository
 
 
 class FileRepositoryFactory:
 
     @staticmethod
-    def get() -> FileRepository:
-        return FileRepository()
+    def get_local_file() -> FileRepository:
+        return LocalFileRepo()
+    
+    @staticmethod
+    def get_s3_file() -> FileRepository:
+        return S3FileRepo()
     
     
 class BiasMetricsFactory:
