+++ backend/core/use_cases/process_link_interactor.py
@@ -1,8 +1,8 @@
-from backend.core.data_access.file_repository import FileRepository
+from backend.core.use_cases.interfaces import FileRepository
 
 class ProcessLinkInteractor:
     def __init__(self, file_repo: FileRepository):
         self.file_repo = file_repo
 
     def post(self, response):
-        return self.file_repo.retrieve_s3_file(response)
+        return self.file_repo.save_file(response)
